Define the PythonBytes type.
This type encapsulates the Python byte string object. This requires a
new data type, mostly because there is no natural "bucket of bytes" type
in Swift without adding a Foundation dependency, which this module
currently does not have.

Instead, we can define a little wrapper that makes it as easy as
possible to convert from existing types to the Python bytes object, and
back again.